-changed the value of claw to an int in ResponseMessage, Controller, and InformationHandler Controller: -changed errorlog to messagelog -all messages sent and received displayed in the MessageLog -tried to do scrolling capabilities with MessageLog. TextArea need to be made to expand with additional messages GraphicsInterface: -added speed TextArea -modified extract to better predict orientation

+++ ControlStation/src/control/communication/ResponseMessage.java
@@ -202,7 +202,7 @@ public class ResponseMessage extends Message
 					touching = true;
 				}
 				int endClaw = responseString.indexOf(":", endTouch + 1);
-				float claw = Float.parseFloat(responseString.substring(
+				int claw = Integer.parseInt(responseString.substring(
 						endClaw + 1, responseString.indexOf("&", endClaw)));
 				int endHeading = responseString.indexOf(":", endClaw + 1);
 				int heading = Integer
@@ -252,7 +252,7 @@ public class ResponseMessage extends Message
 					touching = true;
 				}
 				int endClaw = responseString.indexOf(":", endTouch + 1);
-				float claw = Float.parseFloat(responseString.substring(
+				int claw = Integer.parseInt(responseString.substring(
 						endClaw + 1, responseString.indexOf("&", endClaw)));
 				int endHeading = responseString.indexOf(":", endClaw + 1);
 				int heading = Integer
